checkpoint: integrate Redis 7.2.14 Windows MinGW port
Port the exact Redis 7.2.14 behavior baseline while preserving this fork's Windows IOCP, FDAPI, QFork/jemalloc, service, Event Log, console, and MinGW64 layers.

Core and portability changes:

- adapt Redis 7.2 persistence, multi-part AOF, replication, Cluster, Sentinel, ACL, tracking, networking, and module API behavior to the Windows port
- harden AcceptEx and ConnectEx socket-context updates, completion-error propagation, descriptor lifetime, accepted-socket scheduling, one-shot event rearming, and bounded writable backpressure retries
- update QFork child state, heap/page tracking, persistence callbacks, command-line handling, and Windows-only API restrictions for the 7.2 execution model
- add Windows TLS event rearming and diagnostics while keeping the standard MinGW package explicitly non-TLS

Build, tests, and distribution changes:

- replace active PowerShell process polling and hidden launches in the Tcl harness with a native ownership-checking MinGW launcher
- add deterministic Windows IOCP, AOF, QFork, process-control, replication, Cluster, Sentinel, CLI, and module regressions
- update the MinGW build/module inventory, release metadata, package manifest, Sentinel configuration, Windows documentation, and best-effort Visual Studio/MSI projects

Validation completed for this checkpoint:

- MinGW TLS syntax compile with USE_OPENSSL=1 and BUILD_TLS_MODULE=0
- ./build-mingw.sh -j2
- windows/iocp focused suite: passed in 19 seconds
- windows/regression focused suite: passed in 29 seconds
- Tcl completeness for 32 changed/new scripts and git diff --check

Full clean-build, auxiliary runtime/module, root, module API, Cluster, Sentinel, package extraction, 600-second QFork soak, service restart/EventLog, and PE/configuration/hygiene gates remain pending. The legacy Visual Studio/MSI files are maintained best-effort and are not built in this environment.

+++ tests/modules/aclcheck.c
@@ -38,7 +38,7 @@ int set_aclcheck_key(RedisModuleCtx *ctx, RedisModuleString **argv, int argc) {
         return REDISMODULE_OK;
     }
 
-    RedisModuleCallReply *rep = RedisModule_Call(ctx, "SET", "v", argv + 2, argc - 2);
+    RedisModuleCallReply *rep = RedisModule_Call(ctx, "SET", "v", argv + 2, (size_t)argc - 2);
     if (!rep) {
         RedisModule_ReplyWithError(ctx, "NULL reply returned");
     } else {
@@ -68,7 +68,7 @@ int publish_aclcheck_channel(RedisModuleCtx *ctx, RedisModuleString **argv, int 
         return REDISMODULE_OK;
     }
 
-    RedisModuleCallReply *rep = RedisModule_Call(ctx, "PUBLISH", "v", argv + 1, argc - 1);
+    RedisModuleCallReply *rep = RedisModule_Call(ctx, "PUBLISH", "v", argv + 1, (size_t)argc - 1);
     if (!rep) {
         RedisModule_ReplyWithError(ctx, "NULL reply returned");
     } else {
@@ -98,7 +98,7 @@ int rm_call_aclcheck_cmd(RedisModuleCtx *ctx, RedisModuleUser *user, RedisModule
 
     const char* cmd = RedisModule_StringPtrLen(argv[1], NULL);
 
-    RedisModuleCallReply* rep = RedisModule_Call(ctx, cmd, "v", argv + 2, argc - 2);
+    RedisModuleCallReply* rep = RedisModule_Call(ctx, cmd, "v", argv + 2, (size_t)argc - 2);
     if(!rep){
         RedisModule_ReplyWithError(ctx, "NULL reply returned");
     }else{
@@ -146,7 +146,7 @@ int rm_call_aclcheck_with_errors(RedisModuleCtx *ctx, RedisModuleString **argv, 
 
     const char* cmd = RedisModule_StringPtrLen(argv[1], NULL);
 
-    RedisModuleCallReply* rep = RedisModule_Call(ctx, cmd, "vEC", argv + 2, argc - 2);
+    RedisModuleCallReply* rep = RedisModule_Call(ctx, cmd, "vEC", argv + 2, (size_t)argc - 2);
     RedisModule_ReplyWithCallReply(ctx, rep);
     RedisModule_FreeCallReply(rep);
     return REDISMODULE_OK;
@@ -163,7 +163,7 @@ int rm_call_aclcheck(RedisModuleCtx *ctx, RedisModuleString **argv, int argc){
 
     const char* cmd = RedisModule_StringPtrLen(argv[1], NULL);
 
-    RedisModuleCallReply* rep = RedisModule_Call(ctx, cmd, "vC", argv + 2, argc - 2);
+    RedisModuleCallReply* rep = RedisModule_Call(ctx, cmd, "vC", argv + 2, (size_t)argc - 2);
     if(!rep) {
         char err[100];
         switch (errno) {
